sched: Create SDTL_SKIP flag to skip topology level
If the system admin does not want to use cluster scheduling,
we want to later allow cluster topology level to be
skipped when building sched domains.

Create SDTL_SKIP bit on the sched_domain_topology_level
flag so we can check if a topology level should be skipped
when building sched domains.