riscv: dts: starfive: jh7110-common: revised device node status
The JH7110 is an SoC designed to be suitable for a wide range of
platforms. As new devices are developed based on the JH7110, various
configurations can be anticipated. It's required to leave the
flexibility of enabling device nodes in platform-specific .dts files
instead of within the jh7110-common.dtsi file.

In this patch, remove the status = "okay" lines for devices which are
not "common" in all platforms based on JH7110.

Accordingly, the removed lines are added into those platform-specific
.dts files, including:
    jh7110-milkv-mars.dts
    jh7110-pine64-star64.dts
    jh7110-starfive-visionfive-2.dtsi

Signed-off-by: Alex Elder <elder@riscstar.com>
Signed-off-by: Guodong Xu <guodong@riscstar.com>