spi: k1x: fix build
We want to build a single "spi-k1x" kernel module from two source files,
so we have to tell Kbuild that.  Rename "spi-k1x.c" to be "spi-k1x-main.c"
and update the rules in the Makefile to indicate "spi-k1x" is a single
module made up of "spi-k1x-main" and "spi-k1x-dma".

Also, in "spi-k1x-main.c", tx_hi_thres is never actually used.

Signed-off-by: Alex Elder <elder@riscstar.com>