pwm: Kconfig: add depends on ARCH_SPACEMIT to PWM_PXA
The SpacemiT K1 SoC uses PWM_PXA driver.

This patch updates the Kconfig file for the PWM_PXA driver to add
a dependency on the Spacemit architecture (ARCH_SPACEMIT). This ensures
that the driver can be built for platforms that support the Spacemit
K1 SoC.

With this modification, the PWM_PXA driver can be selected when
ARCH_SPACEMIT is enabled.

Signed-off-by: Guodong Xu <guodong@riscstar.com>