dt-bindings: pwm: marvell,pxa: add optional property resets
This patch updates the device tree bindings for the PXA PWM driver
to add an optional reset property. It modifies the marvell,pxa-pwm.yaml
file to include the resets property, which can reference a reset
controller for the PWM unit, with a maximum of one reset specified.

Added the reset property with a maxItems: 1 constraint, indicating
that only one reset source can be specified.

Signed-off-by: Guodong Xu <guodong@riscstar.com>