riscv: dts: spacemit: Add PWM14 backlight support for BPI-F3
This patch adds a PWM-based backlight node for the Banana Pi BPI-F3 board,
using PWM14. The backlight is defined as a `pwm-backlight` device with
brightness levels and a default brightness setting.

Changes include:
- Defining the `lcd_backlight` node with `pwm-backlight` compatibility
- Assigning PWM14 with a 2kHz frequency for backlight control
- Adding a range of brightness levels from 0 to 255
- Setting `status = "disabled"` to defer enabling until the display
  driver is ready
- Configuring the pinctrl settings for PWM14

This configuration was used to verify PWM driver changes. PWM14 was tested,
and the measured output waveform was correct. The BPI-F3 board uses PWM14
for backlight control, which should be enabled when the display driver
is ready.

Signed-off-by: Guodong Xu <guodong@riscstar.com>