dt-bindings: dma: marvell,mmp-dma: Add support for SpacemiT PDMA
This patch updates the Marvell MMP DMA device tree bindings to add
support for the "spacemit,pdma-1.0" compatible string, allowing the
SpacemiT PDMA controller to be properly recognized and configured
in the device tree.

It also adds optional clocks and resets properties, which may be
required for SpacemiT PDMA integration depending on the platform.

This and the following 14 patches come from this branch from Guodong:
  working_dma_v6.15-rc5_0507_finalizing

Signed-off-by: Guodong Xu <guodong@riscstar.com>