mmc: sdhci-of-k1: use readw() for 16-bit values
In spacemit_sdhci_set_uhs_signaling() there is a case where some bits
should be set in the SDHCI_HOST_CONTROL2 register.  This is donne using
spacemit_sdhci_setbits(), but that does 32-bit accesses (with readl()
and writel()), while that register is 16 bits wide (and 16-bit aligned).

For eMMC (which is all we claim to support right now), the "no-sdio"
property should be set, which means this block of conditional code
will not be executed.

This bug fix has not been submitted for upstream review.

Signed-off-by: Alex Elder <elder@riscstar.com>