clk: spacemit: mark pll1_d8 as critical
This is a quick workaround that prevents pll1_d8 from being disabled
accidentally.  There might be a way to have the clock be marked as
"initially on, enabled by the boot loader".  But for now this makes
things work until all the actual users of this clock are getting
references to it.

Fixes: 1b72c59db0add ("clk: spacemit: Add clock support for SpacemiT K1 SoC")
Signed-off-by: Alex Elder <elder@riscstar.com>