spi: spacemit: introduce k1_spi_set_cs()
Introduce a new function to manipulate chip select.  This function
was obscured by the field name HOLD_FRAME_LOW, but Mark Brown
recognized that's what it was doing, and suggested implementing
the set_cs() callback.  The new function follows that prototype,
but it isn't set up as a callback yet.

Call the function to set chip select in k1_spi_setup(), separating the
setting of the HOLD_FRAME_LOW bit from the rest of the configuration.

Call it also at the beginning and end of k1_spi_transfer_one_message().

Signed-off-by: Alex Elder <elder@riscstar.com>