spi: spacemit: have k1_spi_transfer_one() return positive
When a transfer is successful, have k1_spi_transfer_one() return
a positive value rather than zero.  Don't wait for the transfer
to complete unless a positive value is returned.

If the wait returns a negative value, record it in the message
status; that will be checked afterward and will break out of
the loop if it has been set.

Signed-off-by: Alex Elder <elder@riscstar.com>