spi: spacemit: don't clear message->status
The core code sets message->status to -EINPROGRESS when processing
it is underway.  If it is ever set to anything different, it shows
the message should either end in error, or is successful.

Stop zeroing message->status in k1_spi_transfer_one_message(), and
let it get zeroed (by assignment of "ret") after the entire message
has been transferred.

If the status is still -EINPROGRESS after all transfers have
completed (or an error has occurred), set the message status
to whatever ret is (0/success or error).

Signed-off-by: Alex Elder <elder@riscstar.com>