spi: spacemit: record current transfer
Record the current transfer pointer once we know we'll be starting
it in k1_spi_transfer_one().  This will be updated each time we
start a new transfer; reset it to NULL once we're done with all
message transfers (including if we end early due to error).

If an error is detected in the interrupt handler, set the
SPI_TRANS_FAIL_IO bit in the current transfer's error field
rather than setting the message status.

Have k1_spi_transfer_wait() then return -EIO if it discovers
that bit is set in the current transfer that is completing.

Signed-off-by: Alex Elder <elder@riscstar.com>