spi: spacemit: RX done implies TX done
SPI protocol always sends and receives the same number of bytes for
any transfer.  Bytes are sent by the master before they are received.

This means that RX cannot be done unless TX is also done.  So we
if we find RX is done, we can concluded both TX and RX are done.

Do some general cleanup/refactoring of the handler.

Signed-off-by: Alex Elder <elder@riscstar.com>