spi: spacemit: near-final interrupt handler refactoring
General cleanup and reorganization.  The major thing is to have
the error case share code at the end of the function with the
normal completion case.  The difference is that before completing
the transfer the SPI_TRANS_FAIL_IO bit is set.

We should probably somehow clear out the RX FIFO on error, but
maybe it's just easier to do that before each transfer?  (I'll
look at cleaning it out at the top level if an error occurs.)

Signed-off-by: Alex Elder <elder@riscstar.com>