spi: spacemit: reorder a few things
Set the word size for a transfer separate from enabling the port.

Defer clearing interrupt status until right before interrupts
are enabled.  Also, just write ~0 to clear them rather than
bothering to read.

Rearrange what happens when a transfer is done to be the reverse
of the order these things are set up.

Signed-off-by: Alex Elder <elder@riscstar.com>