spi: spacemit: len is bytes, not words
Currently, spi_transfer_one() assumes that the transfer->len field
is in words, but it's actually in bytes.  We need to divide len
by the bytes need to represent the number of bits per word to
know how many words to read or write.

Move the assignment of values from the transfer field later,
after the bytes per word has been calculated, but also after
the clock rate has been set.

Signed-off-by: Alex Elder <elder@riscstar.com>