spi: spacemit: use the transfer buffer pointers
Stop recording copies of the the transfer rx_buf and tx_buf, and
instead just use the pointers in the transfer.  Track only the
rx_resid and tx_resid values, and compute the offset for reading
or writing the transfer buffers based on that and the length.

Reorganize k1_spi_read_word() and k1_spi_write_word() so they
are structurally very similar.

Signed-off-by: Alex Elder <elder@riscstar.com>