riscv: Add ISA extension parsing for Supm
TODO: drop the change in Kconfig

Supm has been ratified in the RISC-V Pointer Masking specification
(Version 1.0, 10/2024) and is mandated in RVA23 Profiles (Version 1.0,
2024-10-17) for RVA23U64. Supm indicates userspace pointer masking
support.

Add Supm as a standalone extension in the ISA parsing logic. Implement a
validator to ensure that Supm is only reported as available if the
underlying Ssnpm (for supervisor mode) or Smnpm (for machine mode)
extension is present, as Supm relies on them for functionality.

This change also removes the previous macro aliasing of Supm to
Ssnpm/Smnpm in hwcap.h, treating it as a distinct extension ID.

Update the help text of the RISCV_ISA_SUPM Kconfig option to reflect
this new logic and clarify the behavior when the option is enabled.

Signed-off-by: Guodong Xu <guodong@riscstar.com>